Refactor starmap and harvest ownership
