+++ docs/TARGET.md
@@ -2384,6 +2384,8 @@ The target `src/native_app/**` module map is:
 * `src/native_app/app/**` owns native-app state shape, messages, update routing, startup-loaded state, and orchestration between product domains. It may connect app chrome, sample-library behavior, waveform behavior, metadata, audio, shell lifecycle, and workflows, but it should not become a rendering dumping ground or hide broad domain APIs behind convenience re-exports.
 * `src/native_app/app_chrome/**` owns Wavecrate window composition and product-specific view rendering: top and bottom chrome, toolbar, status bars, center-panel layout, source/folder sidebar placement, sample-list panel placement, waveform panel placement, tag-library panel placement, modals, popovers, context-menu overlays, drag previews, and other visible shell layers. It may adapt product state into Radiant views, but it should keep inputs narrow and should not own source scanning, file mutation, metadata persistence, playback, waveform editing rules, or sample-library command side effects.
 * `src/native_app/sample_library/**` owns source, folder, and sample state; source scanning; source watching; file rows; collections; ratings; folder and file commands; drag/drop domain handling; reveal/copy/move/trash workflows; and sample-library context targets. Behavior belongs in this module unless a change is explicitly splitting view code into `app_chrome`.
+  * `folder_browser/starmap/**` owns browser-source layout requests, cached layout and preview-warm indexes, similarity-aware result shaping, and browser selection navigation. It must remain independent from the Starmap rendering composition under `app_chrome`.
+  * `harvest_tracking/**` owns Harvest identity and family queries, background persistence handoff, derivation reconciliation, destination routing, relationship navigation, target selection, and state mutation. Database and filesystem work must stay in persistence or background-task paths rather than view construction or synchronous UI/update handlers.
 * `src/native_app/waveform/**` owns waveform state, audio-derived waveform data, viewport state, play/edit selection state, waveform interaction rules, waveform widget props, waveform cache helpers, and waveform-domain tests. `app_chrome::waveform_panel` may frame and paint this state, but waveform state and editing/audition rules stay in `waveform`.
 * `src/native_app/metadata/**` owns tag vocabulary, tag categories, tag completion data, metadata assignment rules, display-category projection, metadata persistence, and metadata tests. Chrome may paint tag panels and tag pills, but metadata rules and persistence stay here.
 * `src/native_app/audio/**` owns playback, output-device settings, sample loading, cache warming, normalization actions, and audio progress behavior. Chrome may expose controls and status, but audio state transitions and playback policy stay here.
